+++ crates/contracts/tests/asset_auth.rs
@@ -1,0 +1,2 @@
+#[path = "asset_auth/mod.rs"]
+mod asset_auth_tests;
