+++ crates/contracts/tests/asset_auth/mod.rs
@@ -1,0 +1,5 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod burn;
+mod happy_path;
