+++ crates/contracts/tests/common/mod.rs
@@ -1,0 +1,4 @@
+pub mod asserts;
+pub mod issuance;
+pub mod tx_steps;
+pub mod wallet;
