+++ crates/contracts/tests/lending.rs
@@ -1,0 +1,2 @@
+#[path = "lending/mod.rs"]
+mod lending_tests;
