+++ crates/contracts/tests/lending/mod.rs
@@ -1,0 +1,5 @@
+#[test]
+fn lending_module_placeholder() {
+    // Reserved for dedicated lending integration scenarios.
+    assert!(true);
+}
