+++ crates/contracts/tests/pre_lock.rs
@@ -1,0 +1,2 @@
+#[path = "pre_lock/mod.rs"]
+mod pre_lock_tests;
