+++ crates/contracts/tests/pre_lock/cancel.rs
@@ -1,0 +1,14 @@
+use super::support::{cancel_pre_lock_tx, setup_pre_lock};
+
+#[simplex::test]
+fn cancels_pre_lock(context: simplex::TestContext) -> anyhow::Result<()> {
+    let provider = context.get_provider();
+
+    let (txid, pre_lock) = setup_pre_lock(&context)?;
+    provider.wait(&txid)?;
+
+    let txid = cancel_pre_lock_tx(&context, pre_lock, txid)?;
+    provider.wait(&txid)?;
+
+    Ok(())
+}
