+++ crates/contracts/tests/pre_lock/mod.rs
@@ -1,0 +1,7 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod cancel;
+mod create;
+mod create_lending;
+mod support;
