+++ crates/contracts/tests/script_auth.rs
@@ -1,0 +1,2 @@
+#[path = "script_auth/mod.rs"]
+mod script_auth_tests;
