+++ crates/contracts/tests/script_auth/mod.rs
@@ -1,0 +1,4 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod happy_path;
