--- crates/contracts/tests/utils/mod.rs
@@ -1,7 +1,0 @@
-pub mod asset_helpers;
-pub mod signer;
-pub mod split_helpers;
-
-pub use asset_helpers::*;
-pub use signer::*;
-pub use split_helpers::*;
